Use raw pointer casts for slice, str's .len()
This removes the uses of Repr::repr() from libcore. We use a more direct
approach by casting to `Slice<T>` directly. This reduces the
functions and llvm IR used to compile .len(), which should be worth it,
since it is used very often.

With this change, I see compile time improvements for release mode
when compiling slice and indexing heavy code.