Wip: Use unchecked next in .chars() utf-8 decoding
The result.

Taken from: nom project. Quick port of their http request parser to
use `char`. (For no reason, it's slower that way).

The see what the impact of using this change to .chars() is

```
 name                    real-before-3 ns/iter  before-chars-1 ns/iter  diff ns/iter   diff %
 multiple_requests       81,568                 83,306                         1,738    2.13%
 single_request          1,239                  1,235                             -4   -0.32%
 single_request_large    1,895                  2,064                            169    8.92%
 single_request_minimal  329                    294                              -35  -10.64%
```

My guess is that: These benchmarks are just ascii, so the non-ascii
cases's code getting smaller doesn't benefit it, instead it may even get
in the way.